% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_intron_metrics.R
\name{calculate_intron_metrics}
\alias{calculate_intron_metrics}
\title{Calculate Intron Metrics}
\usage{
calculate_intron_metrics(
  mode = "coordinate",
  chromosome = NULL,
  start = NULL,
  end = NULL,
  strand = NULL,
  sequence = NULL
)
}
\arguments{
\item{mode}{Character string specifying the mode of operation. Use "coordinate" to fetch the sequence
from the genome using chromosome coordinates, or "sequence" to provide the sequence directly.
Default is "coordinate".}

\item{chromosome}{Character string specifying the chromosome (e.g., "chr1"). Required if mode is "coordinate".}

\item{start}{Integer specifying the start position of the intron on the chromosome. Required if mode is "coordinate".}

\item{end}{Integer specifying the end position of the intron on the chromosome. Required if mode is "coordinate".}

\item{strand}{Character string specifying the strand ("+" or "-"). Required in both modes.}

\item{sequence}{Character string specifying the intron sequence. Required if mode is "sequence".}
}
\value{
A data frame containing the calculated metrics for the intron:
\item{chromosome}{Chromosome name}
\item{start}{Start position of the intron}
\item{end}{End position of the intron}
\item{strand}{Strand information}
\item{intron_length}{Length of the intron}
\item{percentage_cg}{Percentage of CG content in the intron}
\item{count_polyP}{Count of polyP (C and T) tracts}
\item{polyp_sequence}{Sequence of the polyP tract}
\item{intron_start_motif}{First 10 nucleotides of the intron}
\item{intron_end_motif}{Last 10 nucleotides of the intron}
\item{ESS_region}{ESS region sequence (only for "coordinate" mode)}
\item{ESE_region}{ESE region sequence (only for "coordinate" mode)}
\item{sequence}{Full intron sequence}
}
\description{
This function calculates various metrics for a given intron sequence. It can operate in two modes:
"coordinate" mode, where the user provides chromosome coordinates to fetch the sequence, and
"sequence" mode, where the user directly provides the intron sequence. If it is using the "sequence" mode ESE
and ESS extraction do not work. This also requires a genome that the user selected before everything.
}
\examples{
#' # Example using chromosome coordinates
calculate_intron_metrics(mode = "coordinate", chromosome = "chr1", start = 100000, end = 100100, strand = "-")

# Example using direct sequence input
calculate_intron_metrics(mode = "sequence", sequence = "AGCTTGCGATCGTAGCTAGCTAGCTGACTGACTGACTGACTGACTGACTGAC", strand = "-")

}
