% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_IR_reference_events.R
\name{find_IR_reference_events}
\alias{find_IR_reference_events}
\title{Process GFF File and Identify Unique Exons with Wiggle Room}
\usage{
find_IR_reference_events(gff_file, fa_file, wiggle = 10)
}
\arguments{
\item{gff_file}{A string specifying the path to the GFF file. The file should match the species genome version used in other steps.}

\item{fa_file}{A string specifying the path to the transcript-to-gene mapping file in FASTA format should match the verison of the gff..}

\item{wiggle}{An integer specifying the wiggle room in nucleotides for comparing exon positions. Default is 10.}
}
\value{
A data frame containing unique exons identified in the retained intron transcripts, considering the specified wiggle room.
}
\description{
This function processes a GFF file to filter the species exon data, joins it with transcript-to-gene mapping data,
and identifies the exons to be used for comparison considering a user-defined wiggle room
which depends on annotation. Sugessted wiggle room for human data is 10. This takes considerable time to run
which depends on the size of your gff file, Should be ran only once per analysis
}
\examples{
gff_file <- "ref/gencode.v38.chr_patch_hapl_scaff.annotation.gff3.gz"
fa_file <- "ref/gencode.v38.transcripts.fa.gz"
wiggle <- 10
unique_exons <- process_gff_and_find_unique_exons(gff_file, fa_file, wiggle)
print(unique_exons)
}
